requirements.txt
asgiref==3.4.1
black==21.9b0
boto3==1.18.51
botocore==1.21.51
cffi==1.15.0
click==8.0.1
cryptography==36.0.0
Django==3.2.7
django-cors-headers==3.9.0
django-filter==21.1
django-rest-knox==4.1.0
django-storages==1.11.1
django-tinymce==3.3.0
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.0
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
regex==2021.9.30
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
